feat: add ControllerMixin
